requirements.txt
pymongo==4.6.2
python-decouple==3.8
"pymongo[srv]"
